accountController aupdated and create pocket and transfer to pocket done and tested on postman

+++ jeliBank-Backend/src/main/java/com/example/jeliBankBackend/repository/PocketRepository.java
@@ -3,5 +3,5 @@ package com.example.jeliBankBackend.repository;
 import com.example.jeliBankBackend.model.Pocket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PocketRepository extends JpaRepository<Pocket, Long> {
+public interface PocketRepository extends JpaRepository<Pocket, Integer> {
 }
